Add files via upload
State Contiguity: 
------------------
When studying state interaction, researchers often want to account for spatial effects. A crisis in one state, for instance, is more likely to affect a neighboring state than merely a state in the same hemisphere or continent. Similarly, if we believe that any benefits given to a state also help a neighboring state via spillovers, then not accounting for the contiguity can undermine our assumption of independence across observations. Spatial effects are complicated by the fact that new states emerge over time by gaining independence, or the collapse of a larger empire - that is, any useful contiguity matrix, by necessity must also have a time-series component. 

In this particular case, I wanted to calculate how much foreign aid a given donor provides globally, how much the donor gives to all countries except for the target country, and how much aid is given to the world except for the target country and its neighbors. With this in mind, I set out to create such a contiguity matrix.

Note: The original data on membership in the system of states is from the 2016 Correlates of War dataset. State contiguity is drawn from the directed dyad dataset on contiguity (version 3.2.0) also from Correlates of War. Aiddata.org is a wonderful resource for data on foreign aid, but for the purposes of this coding example, I merely simulated foreign aid flows between dyads rather than using real data.