Add files via upload
The files contained here are the replication files to a project on home country attributes and their relations with potential destinations for foreign direct investment. The project was co-authored with Nate Jensen (UT Austin), but I bear full responsibility for the code. The data management was conducted in Stata, the analysis that followed was done in R.